Refactoring of fragments * collect fragments in package
