+++ app/src/main/res/layout/fragment_home.xml
@@ -5,7 +5,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".SecondFragment">
+    tools:context=".home.HomeFragment">
 
     <TextView
         android:id="@+id/textview_second"
