+++ app/src/main/res/navigation/nav_graph.xml
@@ -7,12 +7,12 @@
 
     <fragment
         android:id="@+id/FirstFragment"
-        android:name="com.example.smarteden.LoginFragment"
-        android:label="@string/first_fragment_label"
+        android:name="com.example.smarteden.login.LoginFragment"
+        android:label="@string/login_fragment_label"
         tools:layout="@layout/fragment_login"/>
     <fragment
         android:id="@+id/SecondFragment"
-        android:name="com.example.smarteden.SecondFragment"
-        android:label="@string/second_fragment_label"
-        tools:layout="@layout/fragment_second"/>
+        android:name="com.example.smarteden.home.HomeFragment"
+        android:label="@string/home_fragment_label"
+        tools:layout="@layout/fragment_home"/>
 </navigation>
