+++ app/src/main/res/values/strings.xml
@@ -2,14 +2,9 @@
     <string name="app_name">Smart EDEN</string>
     <string name="action_settings">Settings</string>
     <!-- Strings used for fragments for navigation -->
-    <string name="first_fragment_label">Login</string>
-    <string name="second_fragment_label">Home</string>
+    <string name="login_fragment_label">Login</string>
+    <string name="home_fragment_label">Home</string>
     <string name="next">Next</string>
     <string name="previous">Previous</string>
-
-    <string name="hello_first_fragment">Hello first fragment</string>
-    <string name="hello_second_fragment">Hello second fragment. Arg: %1$s</string>
-    <string name="your_web_client_id">725204801131-2gd4nc78cb8e45dc2705i2tm5fkj2pia.apps.googleusercontent.com</string>
-    <string name="google_sign_in">Sign in</string>
 
 </resources>
